pyproject.toml
[project]
name = "plugins-adapter"
version = "0.1.0"
description = "Envoy ext-proc for configuring and invoking guardrails in an Envoy-based gateway"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "betterproto2==0.9.1",
    "cpex==0.1.0.dev10",
]

[dependency-groups]
proto = [
    "requests==2.33.0",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "betterproto2==0.9.1",
    "betterproto2_compiler==0.9.0",
    "structlog==25.5.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio==1.3.0",
    "pre-commit>=4.2.1",
]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [
    ".venv",
    "src/envoy",
    "src/udpa",
    "src/validate",
    "src/xds",
]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.isort]
known-first-party = ["src", "tests"]
known-third-party = ["cpex", "envoy", "grpc", "google"]

[tool.pytest.ini_options]
log_cli = false
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(module)s] [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_level = "INFO"
log_format = "%(asctime)s [%(module)s] [%(levelname)s] %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = ["tests"]
pythonpath = [".", "src"]
markers = [
    "integration: integration tests (start real gRPC server)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
